Move the maintenance factories to model package
